pyproject.toml
[tool.pytest.ini_options]
addopts = "--cov-report xml:cov.xml"
norecursedirs = [
    ".venv",
    ".vscode",
    "__pycache__"
]
python_files = [
    "test_*.py"
]
python_functions = [
    "test_*"
]

[tool.coverage.run]
omit = [
    ".venv/*",
    "tests/*",
    "asgi.py"
]

[tool.ruff]
# Same as Black.
line-length = 120
# Assume Python 3.11.
target-version = "py311"

[tool.ruff.lint]
extend-select = ["I001", "I002"]
exclude = [
    ".git",
    ".git-rewrite",
    ".mypy_cache",
    ".ruff_cache",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist"
]

fixable = ["I001", "F401", "W291", "W292"]
# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.lint.per-file-ignores]
"poetry.lock" = ["E501"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
docstring-code-format = true

[tool.isort]
profile = "black"

[tool.poetry]
name = "ytdl-api"
version = "1.8.2"
description = "API for web-based youtube-dl client"
authors = ["Nazar Oleksiuk <nazarii.oleksiuk@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.1"
uvicorn = "^0.18.2"
sse-starlette = "^0.6.1"
aiofiles = "^0.6.0"
ffmpeg-python = "^0.2.0"
confz = "2.0.1"
pyhumps = "^3.7.2"
pytube = "^15.0.0"
yt-dlp = "2024.10.22"
humanize = "^4.8.0"
croniter = "^2.0.5"
tinydb = "^4.8.2"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
pytest-cov = "^6.0.0"
coverage = {extras = ["toml"], version = "^7.6.4"}
pyclean = "^2.2.0"
pytest-mock = "^3.10.0"
httpx = "^0.23.3"
ruff = "^0.7.3"
pytest-asyncio = "^0.23.7"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.masonry.api"
